Refactor list component to be a single file
