--- src/components/ui/list/index.ts
@@ -1,5 +1,0 @@
-export * from './list';
-export * from './list-item';
-export * from './list-item-icon';
-export * from './list-item-text';
-export * from './list-subheader';
