--- src/components/ui/list/list.tsx
@@ -1,72 +1,0 @@
-'use client';
-
-import { type ComponentPropsWithoutRef, forwardRef } from 'react';
-import { RovingFocusGroup } from '@radix-ui/react-roving-focus';
-
-import { cn } from '@/lib/cn';
-
-import {
-  LIST_GROUP_NAME,
-  ListContext,
-  type ScopedProps,
-  useRovingFocusGroupScope,
-} from './common';
-
-type RovingFocusGroupProps = ComponentPropsWithoutRef<typeof RovingFocusGroup>;
-
-export type ListProps = ComponentPropsWithoutRef<'ul'> & {
-  disabled?: boolean;
-  rovingFocus?: boolean;
-  loop?: RovingFocusGroupProps['loop'];
-  dir?: RovingFocusGroupProps['dir'];
-};
-
-// eslint-disable-next-line react/display-name
-const ListUl = forwardRef<HTMLUListElement, ComponentPropsWithoutRef<'ul'>>(
-  ({ className, ...props }, ref) => (
-    <ul
-      {...props}
-      ref={ref}
-      role='group'
-      className={cn('flex w-full flex-col gap-px rounded-lg px-2', className)}
-    />
-  )
-);
-
-export const List = forwardRef<HTMLUListElement, ScopedProps<ListProps>>(
-  (
-    {
-      __scopeList,
-      disabled = false,
-      rovingFocus = true,
-      loop = false,
-      className,
-      ...props
-    },
-    ref
-  ) => {
-    const rovingFocusGroupScope = useRovingFocusGroupScope(__scopeList);
-
-    return (
-      <ListContext
-        scope={__scopeList}
-        rovingFocus={rovingFocus}
-        disabled={disabled}
-      >
-        {rovingFocus ? (
-          <RovingFocusGroup
-            asChild
-            {...rovingFocusGroupScope}
-            orientation='vertical'
-            loop={loop}
-          >
-            <ListUl {...props} ref={ref} className={className} />
-          </RovingFocusGroup>
-        ) : (
-          <ListUl {...props} ref={ref} className={className} />
-        )}
-      </ListContext>
-    );
-  }
-);
-List.displayName = LIST_GROUP_NAME;
